// src/common/open-api/models/IntegrationApi.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Tag } from './Tag';
export type IntegrationApi = {
  api?: string;
  configuration?: Record<string, Record<string, any>>;
  createdBy?: string;
  createdOn?: number;
  description?: string;
  enabled?: boolean;
  integrationName?: string;
  tags?: Array<Tag>;
  updatedBy?: string;
  updatedOn?: number;
};