// src/common/open-api/models/MessageTemplate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Tag } from './Tag';
export type MessageTemplate = {
  createdBy?: string;
  createdOn?: number;
  description?: string;
  integrations?: Array<string>;
  name?: string;
  tags?: Array<Tag>;
  template?: string;
  updatedBy?: string;
  updatedOn?: number;
  variables?: Array<string>;
};